requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
anthropic==0.34.0
PyJWT==2.9.0
httpx==0.27.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-bdd==7.3.0
pytest-asyncio==0.24.0
